+++ ui/app/src/components/ItemDisplay/utils.tsx
@@ -15,58 +15,55 @@
  */
 
 import { ItemStateMap, ItemStates } from '../../models/Item';
-import { FormattedMessage } from 'react-intl';
+import { IntlFormatters } from 'react-intl';
 import * as React from 'react';
 
-export function getItemPublishingTargetText(stateMap: ItemStateMap) {
-	return stateMap.live ? (
-		<FormattedMessage id="words.live" defaultMessage="Live" />
-	) : stateMap.staged ? (
-		<FormattedMessage id="words.staged" defaultMessage="Staged" />
-	) : (
-		<FormattedMessage id="words.unpublished" defaultMessage="Unpublished" />
-	);
+export function getItemPublishingTargetText(
+	stateMap: ItemStateMap,
+	formatMessage: IntlFormatters['formatMessage']
+): string {
+	return stateMap.live
+		? formatMessage({ id: 'words.live', defaultMessage: 'Live' })
+		: stateMap.staged
+			? formatMessage({ id: 'words.staged', defaultMessage: 'Staged' })
+			: formatMessage({ id: 'words.unpublished', defaultMessage: 'Unpublished' });
 }
 
-export function getItemStateText(stateMap: ItemStateMap, values?: Record<string, any>) {
-	let map: { [key in ItemStates]: any };
-	map = {
-		new: () => <FormattedMessage id="itemState.new" defaultMessage="New" />,
-		modified: () => <FormattedMessage id="itemState.modified" defaultMessage="Modified" />,
-		deleted: () => <FormattedMessage id="itemState.deleted" defaultMessage="Deleted" />,
-		locked: () =>
-			values?.user ? (
-				<FormattedMessage id="itemState.lockedBy" defaultMessage="Locked by {user}" values={values} />
-			) : (
-				<FormattedMessage id="itemState.locked" defaultMessage="Locked" />
-			),
-		systemProcessing: () => <FormattedMessage id="itemState.systemProcessing" defaultMessage="System Processing" />,
-		submitted: () => <FormattedMessage id="itemState.submitted" defaultMessage="Submitted" />,
-		scheduled: () => <FormattedMessage id="itemState.scheduled" defaultMessage="Scheduled" />,
-		publishing: () => <FormattedMessage id="itemState.publishing" defaultMessage="Publishing" />,
-		submittedToStaging: () =>
-			stateMap.submitted ? (
-				<FormattedMessage id="itemState.submittedToStaging" defaultMessage="Submitted to staging" />
-			) : (
-				<FormattedMessage id="itemState.scheduledToStaging" defaultMessage="Scheduled to staging" />
-			),
-		submittedToLive: () =>
-			stateMap.submitted ? (
-				<FormattedMessage id="itemState.submittedToLive" defaultMessage="Submitted to live" />
-			) : (
-				<FormattedMessage id="itemState.scheduledToGoLive" defaultMessage="Scheduled to live" />
-			),
-		staged: () => void 0,
-		live: () => void 0,
-		disabled: () => <FormattedMessage id="itemState.disabled" defaultMessage="Disabled" />,
+export function getItemStateText(
+	stateMap: ItemStateMap,
+	formatMessage: IntlFormatters['formatMessage'],
+	values?: { user: string }
+): string {
+	const map: { [key in ItemStates]: string } = {
+		new: formatMessage({ id: 'itemState.new', defaultMessage: 'New' }),
+		modified: formatMessage({ id: 'itemState.modified', defaultMessage: 'Modified' }),
+		deleted: formatMessage({ id: 'itemState.deleted', defaultMessage: 'Deleted' }),
+		locked: values?.user
+			? (formatMessage(
+					{ id: 'itemState.lockedBy', defaultMessage: 'Locked by {user}' },
+					{ user: values.user }
+				) as string)
+			: formatMessage({ id: 'itemState.locked', defaultMessage: 'Locked' }),
+		systemProcessing: formatMessage({ id: 'itemState.systemProcessing', defaultMessage: 'System Processing' }),
+		submitted: formatMessage({ id: 'itemState.submitted', defaultMessage: 'Submitted' }),
+		scheduled: formatMessage({ id: 'itemState.scheduled', defaultMessage: 'Scheduled' }),
+		publishing: formatMessage({ id: 'itemState.publishing', defaultMessage: 'Publishing' }),
+		submittedToStaging: stateMap.submitted
+			? formatMessage({ id: 'itemState.submittedToStaging', defaultMessage: 'Submitted to staging' })
+			: formatMessage({ id: 'itemState.scheduledToStaging', defaultMessage: 'Scheduled to staging' }),
+		submittedToLive: stateMap.submitted
+			? formatMessage({ id: 'itemState.submittedToLive', defaultMessage: 'Submitted to live' })
+			: formatMessage({ id: 'itemState.scheduledToGoLive', defaultMessage: 'Scheduled to live' }),
+		staged: null,
+		live: null,
+		disabled: formatMessage({ id: 'itemState.disabled', defaultMessage: 'Disabled' }),
 		translationUpToDate: null,
 		translationPending: null,
 		translationInProgress: null
 	};
+
 	return (
-		map[getItemStateId(stateMap)]?.() ?? (
-			<FormattedMessage id="itemState.notInWorkflow" defaultMessage="Not in workflow" />
-		)
+		map[getItemStateId(stateMap)] ?? formatMessage({ id: 'itemState.notInWorkflow', defaultMessage: 'Not in workflow' })
 	);
 }